# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Local Music Annotation Service
  version: 1.0.0-draft
  description: >-
    Local sidecar API for score-aligned, human-in-the-loop music annotation.
    All mutating endpoints are auditable. Long operations are asynchronous jobs.
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "49321"
security:
  - sessionToken: []
...
paths:
...
  /v1/providers/{providerId}:load:
    post:
      tags: [Models]
      operationId: loadProvider
      parameters:
        - name: providerId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderLoadRequest" }
      responses:
        "202":
          description: Model load job accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobAccepted" }
  /v1/providers/{providerId}:probe:
    post:
      tags: [Models]
      operationId: probeProvider
      parameters:
        - name: providerId
          in: path
          required: true
          schema: { type: string }
      responses:
        "202":
          description: One-second audio and optional text probe accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobAccepted" }
  /v1/providers/{providerId}:unload:
    post:
      tags: [Models]
      operationId: unloadProvider
      parameters:
        - name: providerId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Provider unloaded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderCapability" }
  /v1/projects:
    post:
      tags: [Projects]
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
      responses:
        "201":
          description: Project created.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      operationId: getProject
      responses:
        "200":
          description: Project summary.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/projects/{projectId}/assets:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Assets]
      operationId: listAssets
      responses:
        "200":
          description: Project assets.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Asset" }
    post:
      tags: [Assets]
      operationId: importAsset
      description: >-
        Import a trusted path resolved by the extension host. The backend hashes
        and inspects the file asynchronously. The Webview must not call this
        endpoint directly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetImport"
      responses:
        "202": { $ref: "#/components/responses/JobAccepted" }
...
  /v1/projects/{projectId}/annotations:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Annotations]
      operationId: queryAnnotations
      parameters:
        - name: assetId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: startSample
          in: query
          schema: { type: integer, minimum: 0 }
        - name: endSample
          in: query
          schema: { type: integer, minimum: 1 }
        - name: state
          in: query
          schema:
            type: array
            items: { $ref: "#/components/schemas/AnnotationState" }
      responses:
        "200":
          description: Annotations overlapping the requested interval.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Annotation" }
    post:
      tags: [Annotations]
      operationId: createAnnotation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnotationCreate" }
      responses:
        "201":
          description: Annotation created.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Annotation" }
  /v1/projects/{projectId}/annotations/{annotationId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - name: annotationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Annotations]
      operationId: updateAnnotation
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/AnnotationPatch" }
      responses:
        "200":
          description: Updated annotation.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Annotation" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Annotations]
      operationId: deleteAnnotation
      description: Soft-deletes the annotation and appends an audit event.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204": { description: Annotation soft-deleted. }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /v1/projects/{projectId}/detection-plans:compile:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      tags: [Detection Plans]
      operationId: compileDetectionPlan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DetectionPlanCompileRequest" }
      responses:
        "202": { $ref: "#/components/responses/JobAccepted" }
  /v1/projects/{projectId}/detection-plans/{planId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - name: planId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Detection Plans]
      operationId: getDetectionPlan
      responses:
        "200":
          description: Versioned detection plan.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "./detection-plan.schema.json" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Detection Plans]
      operationId: updateDetectionPlan
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "./detection-plan.schema.json" }
      responses:
        "200":
          description: Updated plan.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "./detection-plan.schema.json" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/projects/{projectId}/detection-plans/{planId}:validate:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - name: planId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Detection Plans]
      operationId: validateDetectionPlan
      responses:
        "200":
          description: Schema, taxonomy, provider, and calibration validation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetectionPlanValidation" }
  /v1/projects/{projectId}/detection-plans/{planId}:preview:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - name: planId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Detection Plans]
      operationId: previewDetectionPlan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DetectionPlanRunRequest" }
      responses:
        "202": { $ref: "#/components/responses/JobAccepted" }
  /v1/projects/{projectId}/detection-plans/{planId}:run:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - name: planId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Detection Plans]
      operationId: runDetectionPlan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DetectionPlanRunRequest" }
      responses:
        "202": { $ref: "#/components/responses/JobAccepted" }
  /v1/projects/{projectId}/detection-runs/{runId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - name: runId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Detection Plans]
      operationId: getDetectionRun
      responses:
        "200":
          description: Run state and frozen plan revision.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetectionRun" }
  /v1/projects/{projectId}/detection-runs/{runId}/candidates:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - name: runId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Detection Plans]
      operationId: listDetectionCandidates
      parameters:
        - name: assetId
          in: query
          schema: { type: string, format: uuid }
        - name: startSample
          in: query
          schema: { type: integer, minimum: 0 }
        - name: endSample
          in: query
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Candidate labels with evidence and decisions.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DetectionCandidate" }
...
  /v1/jobs/{jobId}:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Jobs]
      operationId: getJob
      responses:
        "200":
          description: Current job state.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/jobs/{jobId}:cancel:
    parameters:
      - name: jobId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Jobs]
      operationId: cancelJob
      responses:
        "202":
          description: Cancellation requested.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  headers:
    ETag:
      description: Quoted resource revision used for optimistic concurrency.
      schema: { type: string, examples: ['"revision-7"'] }
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string }
  responses:
    JobAccepted:
      description: Asynchronous job accepted.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/JobAccepted" }
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: Revision conflict or incompatible project state.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    ProviderCapability:
      type: object
      required: [providerId, installed, supportsFrameEmbeddings, supportsTraining]
      properties:
        providerId: { type: string }
        displayName: { type: string }
        installed: { type: boolean }
        loaded: { type: boolean }
        state: { type: string, enum: [not_installed, available, loading, loaded, failed] }
        supportsFrameEmbeddings: { type: boolean }
        supportsTextEmbeddings: { type: boolean }
        supportsTraining: { type: boolean }
        weightLicense: { type: [string, "null"] }
        commercialUse: { type: [boolean, "null"] }
        sampleRate: { type: integer }
        device: { type: [string, "null"] }
        error: { type: [string, "null"] }
    ProviderLoadRequest:
      type: object
      additionalProperties: false
      properties:
        device: { type: string, enum: [auto, cpu, cuda], default: auto }
        checkpointPath: { type: [string, "null"] }
        allowDownload: { type: boolean, default: false }
    ProjectCreate:
      type: object
      additionalProperties: false
      required: [name, directory]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        directory: { type: string, description: Trusted absolute directory path. }
        description: { type: string, maxLength: 5000, default: "" }
    Project:
      type: object
      required: [id, name, directory, revision, createdAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        name: { type: string }
        directory: { type: string }
        description: { type: string }
        activeTaxonomyVersion: { type: [string, "null"] }
        activeModelVersionId: { type: [string, "null"], format: uuid }
        revision: { type: integer, minimum: 1 }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
...
    AssetImport:
      type: object
      additionalProperties: false
      required: [path]
      properties:
        path: { type: string, description: Trusted absolute audio path resolved by the extension host. }
        name: { type: [string, "null"], maxLength: 500 }
    Asset:
      type: object
      required: [id, name, path, contentHash, sampleRate, channels, samples, durationSeconds, createdAt]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        name: { type: string }
        path: { type: string }
        contentHash: { type: string }
        sampleRate: { type: integer, minimum: 1 }
        channels: { type: integer, minimum: 1 }
        samples: { type: integer, minimum: 1 }
        durationSeconds: { type: number, exclusiveMinimum: 0 }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
...
    AnnotationState:
      type: string
      enum: [suggested, confirmed, rejected, ambiguous]
...
    Annotation:
      allOf:
        - $ref: "#/components/schemas/AnnotationCreate"
        - type: object
          required: [id, revision, createdAt, updatedAt]
          properties:
            id: { $ref: "#/components/schemas/UUID" }
            revision: { type: integer, minimum: 1 }
            createdAt: { $ref: "#/components/schemas/Timestamp" }
            updatedAt: { $ref: "#/components/schemas/Timestamp" }
...
    JobAccepted:
      type: object
      required: [jobId, state]
      properties:
        jobId: { $ref: "#/components/schemas/UUID" }
        state: { $ref: "#/components/schemas/JobState" }
...